chore: add action and renamed tests to test
