+++ src/exceptions.py
@@ -1,6 +1,4 @@
 # Author: Philip De lorenzo <philip.delorenzo@gmail.com>
-
-
 class GHTokenError(Exception):
     def __init__(self, message, errors={}):
         # Call the base class constructor with the parameters it needs
